[#261] use minimal NATS stream storage config
Introduce a dedicated StorageConfig to hold the 
NATS stream storage type and move stream-only 
settings out of the main Settings. Add a small 
storage-only module (nats/storage.py) so stream configs can be 
defined without needing NATS credentials.

Remove bulky NATS mode/credential validation and related fields from
the global config to decouple stream configuration from client
authentication. Update nats/config.py to rely on the separate storage
setting (preparing for simplified stream definitions and avoiding
side effects of credential checks on import).
