+++ backend/callout/test/run.py
@@ -1,7 +1,7 @@
 import asyncio
 
 from core.nats import create_or_update_stream, nc
-from core.nats.config import AGENTS_STREAM_CONFIG
+from core.nats.streams import AGENTS_STREAM_CONFIG
 from nats.aio.client import Client as NATSClient
 from pydantic_settings import BaseSettings, SettingsConfigDict
